pyproject.toml
[tool.poetry]
name = "discord-bot-change-your-nickname-color"
version = "0.1.0"
description = ""
authors = ["LianSheng197 <REDACTED>"]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.2.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
